refactor: rename Radio to RadioButton
